+++ changelog.md
@@ -1,5 +1,12 @@
 # Version log
 ## 0.0.7dev
+### Mon 26 Jun 08:47:56 BST 2023
+- bug fix :  
+  - fixed folder name in `data_path`
+- cosmetic :
+  - `fetch` flag when updating zenodo so it only does it on cloud request
+  - minor display stuff
+  
 ### Wed 21 Jun 14:55:31 BST 2023
 - issue#44 import info:
   - added update zenodo functionality although does not work in headless